Adds module which puts SQS queue in front of a lambda
This adds a module which inserts an SQS queue between an SNS topic and
a lambda. The result is similar to when a lambda reads from SNS, but
the risk of losing messages is mitigated as they're first sent to SQS
and can be put in a dead letter queue on failure.